[#346] Document versioning policy
Problem: Our versioning policy isn't documented anywhere, as a result
users may get confused when they see a package with version that is
different from the upstream release version.

Solution: Write a doc that explains our versioning policy and the way in
which packages for upstream releases and pre-releases are distributed.
Also, refere to this doc in the main README.