[#409] Use formulae tags for attaching bottles to releases
Problem: currently, both on GitHub Actions and with Buildkite, we build brew
bottles on a new tag being pushed. In there runs, the bottle-building CI steps
that have to do with:

1. checking that the bottle has already been built;
2. attaching built bottles to releases;
3. making commits that add new hashes to the formulae;

use the tag that triggered the pipeline. Turns out this behaviour doesn't work
for any subsequent releases we make in the same upstream release, because the
bottle versions usually stay the same, leading e.g. to the v.*-2 release having
v.*-1 bottles attached.

Solution: changed the scripts to use the version of a specific formula where
appropriate, or the one in tezos-client.rb where there is no specific bottle
to be handled, as tezos-client is probably going to use the latest version.
In addition, changed BK steps to also check if a bottle has already been built.