[#517] Use environment files in nodes services
Problem: Currently the environment for node services
is not defined using environment file, but with
several different environment additions to the systemd
service definition.
This doesn't fit well with the other services and is no
longer necessary, since we have config appends now.

Solution: Add env files to node services and move the
necessary defauls and settings there.