[#517] Remove NODE_DATA_DIR from baking services
Problem: the NODE_DATA_DIR env variable is defined
in baking services. However that's a bad place for it
because it causes duplication with the node services,
which means that changes too have to be duplicated.

Solution: removed NODE_DATA_DIR from baking service
definitions and leave it to nodes only.
Additionally remove move RESET_ON_STOP to the custom
node definition.