[#539] Load binaries from release, instead of artifacts
Problem:

We have a bug in the BigSur release pipeline in the following situation:

We a building bottles for BigSur
Some bottles are built, but then exception happens
The build restarts, without building the bottles already built
Only the hashes of the last built bottles are updated.

Solution: In order to prevent this, we take new bottles
from gh release as all bottles built correctly end up there.
This should fix the problem, though it is quite hard
to reproduce it right now