[#718] Improve algorithm of picking snapshot
Problem: Currently setup wizard tries to download the snapshot only
on the same Octez version (see #707) from the provider metadata.
However, this is not correct, since snapshot doesn't have to be created
with the exact same node version.

Solution: Provide one of these latest snapshots in desceasing order of
relevance:
* exact version match
* major version match
* snapshot version is 6