[#738] Weaken the 'snapshot_version' condition
Problem: The current implementation of snapshot
selection algorithm leads to the case when wizard
can't find the compatible snapshot for 'nairobinet'
in xtz-shots provider metadata beceause they are only
snapshot of version 5 for this network.

Solution: Allow 'snapshot_version' field in metadata
to be one less than 'compatible_snapshot_version'.

This change is safe because of the backwards
compatibility of the snapshot versions.