[#793] Add support of stake operation
Problem: In ParisB protocol, each baker needs to explicitly stake some
balance (minimum 600tz) to receive baking rights. In case of the bakers
that are registered in Oxford or earlier, they are staked automatically,
but this is not the case for new bakers.

Solution: Check if explicit staking is relevant for the user,
ask for a stake amount and perform staking.